make service description configurable
The previous commit that changed the service description
from Openstack to OpenStack actually introduced a bug that means
that users cannot upgrade between those two versions of the code.

An attempt to use this manifest to upgrade from earlier versions
of the code before this change was made is currently not supported
because keystone_service does not support updating it's description.

This commit adds the service description as a class parameter so that
folks using older version of the module can update it to the incorrectly
spelled version.

Change-Id: Id8b01c7456b067c19fcf6cc1503c22a11b7d3fd9